fix: Build and test dice-memcpy
The dice-memcpy module implements an interceptor for memset.  To avoid
recursion into that function before the Self guards are set, all object
files in Dice modules should not call memset. Explicit calls to memset
should be replaced with dice_memset and implicit (compiler injected)
calls should be disabled with -fno-builtin-memset.

Signed-off-by: Diogo Behrens <diogo.behrens@huawei.com>